Refactor doctor into shared core
